+++ CHANGELOG.md
@@ -14,22 +14,32 @@
 * Fixed bug.
 
 ## [0.5.0]
+
 * Fixed bug
 * Modified interfaces
 * 1.'headerWidget' => 'headerWidgetBuilder'
 * 2.'footerWidget' => 'footerWidgetBuilder'
 
 ## [0.6.0]
+
 * Added loadMoreWidgetBuilder property.
 
 ## [0.7.0]
+
 * Fixed loadMoreWidgetBuilder property.
 
 ## [0.8.0]
+
 * Removed required property limit.
 
 ## [0.9.0]
+
 * Modified example and screenshot.
 
 ## [1.0.0]
-* Added jumpTo(section,index) and animateTo(section,index,animation) functions. 
+
+* Added jumpTo(section,index) and animateTo(section,index,animation) functions.
+
+## [1.1.0]
+
+* Fixed jumpTo and animateTo function offset bug.